+++ conftest.py
@@ -36,11 +36,16 @@ import numpy as np
 from orix.crystal_map import CrystalMap, Phase, PhaseList, create_coordinate_arrays
 from orix.quaternion import Rotation
 import pytest
-import skimage.color as skc
 
 import kikuchipy as kp
 from kikuchipy import constants
 from kikuchipy.data._data import marshall
+from kikuchipy.data._dummy_files.bruker_h5ebsd import (
+    create_dummy_bruker_h5ebsd_file,
+    create_dummy_bruker_h5ebsd_nonrectangular_roi_file,
+    create_dummy_bruker_h5ebsd_roi_file,
+)
+from kikuchipy.data._dummy_files.oxford_h5ebsd import create_dummy_oxford_h5ebsd_file
 from kikuchipy.io.plugins._h5ebsd import _dict2hdf5group
 
 if constants.installed["pyvista"]:
@@ -50,7 +55,7 @@ if constants.installed["pyvista"]:
     pv.global_theme.interactive = False
 
 
-DATA_PATH = Path(__file__).parent / "src/kikuchipy/data"
+DATA_PATH = Path(kp.data.__file__).parent.resolve()
 
 # ------------------------------ Setup ------------------------------ #
 
@@ -442,11 +447,6 @@ def oxford_binary_path() -> Generator[Path, None, None]:
     yield DATA_PATH / "oxford_binary"
 
 
-@pytest.fixture
-def oxford_h5ebsd_path() -> Generator[Path, None, None]:
-    yield DATA_PATH / "oxford_h5ebsd"
-
-
 @pytest.fixture(params=[((2, 3), (60, 60), np.uint8, 2, False, True)])
 def oxford_binary_file(tmpdir, request) -> Generator[TextIOWrapper, None, None]:
     """Create a dummy Oxford Instruments' binary .ebsp file.
@@ -511,6 +511,9 @@ def oxford_binary_file(tmpdir, request) -> Generator[TextIOWrapper, None, None]:
 
     for i in new_order:
         r, c = np.unravel_index(i, (nr, nc))
+        if ver > 4:
+            extra_pattern_header = np.array([c, r], dtype=np.int32)
+            extra_pattern_header.tofile(f)
         pattern_header.tofile(f)
         data[r, c].tofile(f)
         if ver > 1:
@@ -525,6 +528,13 @@ def oxford_binary_file(tmpdir, request) -> Generator[TextIOWrapper, None, None]:
     f.close()
 
     yield f
+
+
+@pytest.fixture()
+def oxford_h5ebsd_file(tmpdir) -> Generator[TextIOWrapper, None, None]:
+    fpath = tmpdir / "patterns.h5oina"
+    create_dummy_oxford_h5ebsd_file(fpath)
+    yield fpath
 
 
 # -------------------------- EMsoft formats -------------------------- #
@@ -786,115 +796,8 @@ def bruker_path() -> Generator[Path, None, None]:
 @pytest.fixture
 def bruker_h5ebsd_file(tmpdir) -> Generator[Path, None, None]:
     """Bruker h5ebsd file with no region of interest."""
-    s = kp.data.nickel_ebsd_small()
-    ny, nx = s._navigation_shape_rc
-    n = ny * nx
-    sy, sx = s._signal_shape_rc
-
     fpath = tmpdir / "patterns.h5"
-    f = h5py.File(fpath, mode="w")
-
-    # Top group
-    man = f.create_dataset("Manufacturer", shape=(1,), dtype="|S11")
-    man[()] = b"Bruker Nano"
-    ver = f.create_dataset("Version", shape=(1,), dtype="|S10")
-    ver[()] = b"Esprit 2.X"
-    scan = f.create_group("Scan 0")
-
-    # EBSD
-    ebsd = scan.create_group("EBSD")
-
-    ones9 = np.ones(n, dtype=np.float32)
-    zeros9 = np.zeros(n, dtype=np.float32)
-
-    ebsd_data = ebsd.create_group("Data")
-    ebsd_data.create_dataset("DD", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("MAD", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("MADPhase", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("NIndexedBands", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("PCX", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("PCY", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("PHI", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("Phase", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("RadonBandCount", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("RadonQuality", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("RawPatterns", data=s.data.reshape((n, sy, sx)))
-    ebsd_data.create_dataset("X BEAM", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("X SAMPLE", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("Y BEAM", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("Y SAMPLE", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("Z SAMPLE", dtype=np.float32, data=zeros9)
-    ebsd_data.create_dataset("phi1", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("phi2", dtype=np.float32, data=ones9)
-
-    ebsd_header = ebsd.create_group("Header")
-    ebsd_header.create_dataset("CameraTilt", dtype=float, data=0)
-    ebsd_header.create_dataset("DetectorFullHeightMicrons", dtype=np.int32, data=sy)
-    ebsd_header.create_dataset("DetectorFullWidthMicrons", dtype=np.int32, data=sx)
-    grid_type = ebsd_header.create_dataset("Grid Type", shape=(1,), dtype="|S9")
-    grid_type[()] = b"isometric"
-    ebsd_header.create_dataset("KV", dtype=float, data=20)
-    ebsd_header.create_dataset("MADMax", dtype=float, data=1.5)
-    ebsd_header.create_dataset("Magnification", dtype=float, data=200)
-    ebsd_header.create_dataset("MapStepFactor", dtype=float, data=4)
-    ebsd_header.create_dataset("MaxRadonBandCount", dtype=np.int32, data=11)
-    ebsd_header.create_dataset("MinIndexedBands", dtype=np.int32, data=5)
-    ebsd_header.create_dataset("NCOLS", dtype=np.int32, data=nx)
-    ebsd_header.create_dataset("NROWS", dtype=np.int32, data=ny)
-    ebsd_header.create_dataset("NPoints", dtype=np.int32, data=n)
-    original_file = ebsd_header.create_dataset("OriginalFile", shape=(1,), dtype="|S50")
-    original_file[()] = b"/a/home/for/your/data.h5"
-    ebsd_header.create_dataset("PatternHeight", dtype=np.int32, data=sy)
-    ebsd_header.create_dataset("PatternWidth", dtype=np.int32, data=sx)
-    ebsd_header.create_dataset("PixelByteCount", dtype=np.int32, data=1)
-    s_mean = s.nanmean((2, 3)).data.astype(np.uint8)
-    ebsd_header.create_dataset("SEM Image", data=skc.gray2rgb(s_mean))
-    ebsd_header.create_dataset("SEPixelSizeX", dtype=float, data=1)
-    ebsd_header.create_dataset("SEPixelSizeY", dtype=float, data=1)
-    ebsd_header.create_dataset("SampleTilt", dtype=float, data=70)
-    bg = s.static_background
-    ebsd_header.create_dataset("StaticBackground", dtype=np.uint16, data=bg)
-    ebsd_header.create_dataset("TopClip", dtype=float, data=1)
-    ebsd_header.create_dataset("UnClippedPatternHeight", dtype=np.int32, data=sy)
-    ebsd_header.create_dataset("WD", dtype=float, data=1)
-    ebsd_header.create_dataset("XSTEP", dtype=float, data=1.5)
-    ebsd_header.create_dataset("YSTEP", dtype=float, data=1.5)
-    ebsd_header.create_dataset("ZOffset", dtype=float, data=0)
-
-    phase = ebsd_header.create_group("Phases/1")
-    formula = phase.create_dataset("Formula", shape=(1,), dtype="|S2")
-    formula[()] = b"Ni"
-    phase.create_dataset("IT", dtype=np.int32, data=225)
-    phase.create_dataset(
-        "LatticeConstants",
-        dtype=np.float32,
-        data=np.array([3.56, 3.56, 3.56, 90, 90, 90]),
-    )
-    name = phase.create_dataset("Name", shape=(1,), dtype="|S6")
-    name[()] = b"Nickel"
-    phase.create_dataset("Setting", dtype=np.int32, data=1)
-    space_group = phase.create_dataset("SpaceGroup", shape=(1,), dtype="|S5")
-    space_group[()] = b"Fm-3m"
-    atom_pos = phase.create_group("AtomPositions")
-    atom_pos1 = atom_pos.create_dataset("1", shape=(1,), dtype="|S17")
-    atom_pos1[()] = b"Ni,0,0,0,1,0.0035"
-
-    # SEM
-    sem = scan.create_group("SEM")
-    sem.create_dataset("SEM IX", dtype=np.int32, data=np.ones(1))
-    sem.create_dataset("SEM IY", dtype=np.int32, data=np.ones(1))
-    sem.create_dataset("SEM Image", data=skc.gray2rgb(s_mean))
-    sem.create_dataset("SEM ImageHeight", dtype=np.int32, data=3)
-    sem.create_dataset("SEM ImageWidth", dtype=np.int32, data=3)
-    sem.create_dataset("SEM KV", dtype=float, data=20)
-    sem.create_dataset("SEM Magnification", dtype=float, data=200)
-    sem.create_dataset("SEM WD", dtype=float, data=24.5)
-    sem.create_dataset("SEM XResolution", dtype=float, data=1)
-    sem.create_dataset("SEM YResolution", dtype=float, data=1)
-    sem.create_dataset("SEM ZOffset", dtype=float, data=0)
-
-    f.close()
-
+    create_dummy_bruker_h5ebsd_file(fpath)
     yield fpath
 
 
@@ -903,123 +806,9 @@ def bruker_h5ebsd_roi_file(tmpdir) -> Generator[Path, None, None]:
     """Bruker h5ebsd file with rectangular region of interest (and SEM
     group under EBSD group).
     """
-    s = kp.data.nickel_ebsd_small()
-    ny, nx = s._navigation_shape_rc
-    n = ny * nx
-    sy, sx = s._signal_shape_rc
-
-    fpath = tmpdir / "patterns_roi.h5"
-    f = h5py.File(fpath, mode="w")
-
-    # Top group
-    man = f.create_dataset("Manufacturer", shape=(1,), dtype="|S11")
-    man[()] = b"Bruker Nano"
-    ver = f.create_dataset("Version", shape=(1,), dtype="|S10")
-    ver[()] = b"Esprit 2.X"
-    scan = f.create_group("Scan 0")
-
-    # EBSD
-    ebsd = scan.create_group("EBSD")
-
-    # ROI and shape
-    roi = np.array(
-        [
-            [0, 1, 1],  # 0, 1, 2 | (0, 0) (0, 1) (0, 2)
-            [0, 1, 1],  # 3, 4, 5 | (1, 0) (1, 1) (1, 2)
-            [0, 1, 1],  # 6, 7, 8 | (2, 0) (2, 1) (2, 2)
-        ],
-        dtype=bool,
-    ).flatten()
-    # Order of ROI patterns: 4, 1, 2, 5, 7, 8
-    iy = np.array([1, 0, 0, 1, 2, 2], dtype=int)
-    ix = np.array([1, 1, 2, 2, 1, 2], dtype=int)
-
-    # Data
-    ones9 = np.ones(9, dtype=np.float32)[roi]
-    zeros9 = np.zeros(9, dtype=np.float32)[roi]
-    ebsd_data = ebsd.create_group("Data")
-    ebsd_data.create_dataset("DD", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("MAD", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("MADPhase", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("NIndexedBands", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("PCX", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("PCY", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("PHI", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("Phase", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("RadonBandCount", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("RadonQuality", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("RawPatterns", data=s.data.reshape((n, sy, sx))[roi])
-    ebsd_data.create_dataset("X BEAM", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("X SAMPLE", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("Y BEAM", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("Y SAMPLE", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("Z SAMPLE", dtype=np.float32, data=zeros9)
-    ebsd_data.create_dataset("phi1", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("phi2", dtype=np.float32, data=ones9)
-
-    # Header
-    ebsd_header = ebsd.create_group("Header")
-    ebsd_header.create_dataset("CameraTilt", dtype=float, data=0)
-    ebsd_header.create_dataset("DetectorFullHeightMicrons", dtype=np.int32, data=23700)
-    ebsd_header.create_dataset("DetectorFullWidthMicrons", dtype=np.int32, data=31600)
-    grid_type = ebsd_header.create_dataset("Grid Type", shape=(1,), dtype="|S9")
-    grid_type[()] = b"isometric"
-    ebsd_header.create_dataset("KV", dtype=float, data=20)
-    ebsd_header.create_dataset("MADMax", dtype=float, data=1.5)
-    ebsd_header.create_dataset("Magnification", dtype=float, data=200)
-    ebsd_header.create_dataset("MapStepFactor", dtype=float, data=4)
-    ebsd_header.create_dataset("MaxRadonBandCount", dtype=np.int32, data=11)
-    ebsd_header.create_dataset("MinIndexedBands", dtype=np.int32, data=5)
-    ebsd_header.create_dataset("NCOLS", dtype=np.int32, data=nx)
-    ebsd_header.create_dataset("NROWS", dtype=np.int32, data=ny)
-    ebsd_header.create_dataset("NPoints", dtype=np.int32, data=n)
-    original_file = ebsd_header.create_dataset("OriginalFile", shape=(1,), dtype="|S50")
-    original_file[()] = b"/a/home/for/your/data.h5"
-    ebsd_header.create_dataset("PatternHeight", dtype=np.int32, data=sy)
-    ebsd_header.create_dataset("PatternWidth", dtype=np.int32, data=sx)
-    ebsd_header.create_dataset("PixelByteCount", dtype=np.int32, data=1)
-    s_mean = s.nanmean((2, 3)).data.astype(np.uint8)
-    ebsd_header.create_dataset("SEM Image", data=skc.gray2rgb(s_mean))
-    ebsd_header.create_dataset("SEPixelSizeX", dtype=float, data=1)
-    ebsd_header.create_dataset("SEPixelSizeY", dtype=float, data=1)
-    ebsd_header.create_dataset("SampleTilt", dtype=float, data=70)
-    ebsd_header.create_dataset(
-        "StaticBackground", dtype=np.uint16, data=s.static_background
-    )
-    ebsd_header.create_dataset("TopClip", dtype=float, data=1)
-    ebsd_header.create_dataset("UnClippedPatternHeight", dtype=np.int32, data=sy)
-    ebsd_header.create_dataset("WD", dtype=float, data=1)
-    ebsd_header.create_dataset("XSTEP", dtype=float, data=1.5)
-    ebsd_header.create_dataset("YSTEP", dtype=float, data=1.5)
-    ebsd_header.create_dataset("ZOffset", dtype=float, data=0)
-    # Phases
-    phase = ebsd_header.create_group("Phases/1")
-    formula = phase.create_dataset("Formula", shape=(1,), dtype="|S2")
-    formula[()] = b"Ni"
-    phase.create_dataset("IT", dtype=np.int32, data=225)
-    phase.create_dataset(
-        "LatticeConstants",
-        dtype=np.float32,
-        data=np.array([3.56, 3.56, 3.56, 90, 90, 90]),
-    )
-    name = phase.create_dataset("Name", shape=(1,), dtype="|S6")
-    name[()] = b"Nickel"
-    phase.create_dataset("Setting", dtype=np.int32, data=1)
-    space_group = phase.create_dataset("SpaceGroup", shape=(1,), dtype="|S5")
-    space_group[()] = b"Fm-3m"
-    atom_pos = phase.create_group("AtomPositions")
-    atom_pos1 = atom_pos.create_dataset("1", shape=(1,), dtype="|S17")
-    atom_pos1[()] = b"Ni,0,0,0,1,0.0035"
-
-    # SEM
-    sem = ebsd.create_group("SEM")
-    sem.create_dataset("IX", dtype=np.int32, data=ix)
-    sem.create_dataset("IY", dtype=np.int32, data=iy)
-    sem.create_dataset("ZOffset", dtype=float, data=0)
-
-    f.close()
-
-    yield fpath
+    path = tmpdir / "patterns_roi.h5"
+    create_dummy_bruker_h5ebsd_roi_file(path)
+    yield path
 
 
 @pytest.fixture
@@ -1027,120 +816,6 @@ def bruker_h5ebsd_nonrectangular_roi_file(tmpdir) -> Generator[Path, None, None]
     """Bruker h5ebsd file with non-rectangular region of interest (and
     SEM group under EBSD group).
     """
-    s = kp.data.nickel_ebsd_small()
-    ny, nx = s._navigation_shape_rc
-    n = ny * nx
-    sy, sx = s._signal_shape_rc
-
-    fpath = tmpdir / "patterns_roi_nonrectangular.h5"
-    f = h5py.File(fpath, mode="w")
-
-    # Top group
-    man = f.create_dataset("Manufacturer", shape=(1,), dtype="|S11")
-    man[()] = b"Bruker Nano"
-    ver = f.create_dataset("Version", shape=(1,), dtype="|S10")
-    ver[()] = b"Esprit 2.X"
-    scan = f.create_group("Scan 0")
-
-    # EBSD
-    ebsd = scan.create_group("EBSD")
-
-    # ROI and shape
-    roi = np.array(
-        [
-            [0, 1, 1],  # 0, 1, 2 | (0, 0) (0, 1) (0, 2)
-            [0, 1, 1],  # 3, 4, 5 | (1, 0) (1, 1) (1, 2)
-            [0, 1, 0],  # 6, 7, 8 | (2, 0) (2, 1) (2, 2)
-        ],
-        dtype=bool,
-    ).flatten()
-    # Order of ROI patterns: 4, 1, 2, 7, 5
-    iy = np.array([1, 0, 0, 2, 1], dtype=int)
-    ix = np.array([1, 1, 2, 1, 2], dtype=int)
-
-    # Data
-    ones9 = np.ones(n, dtype=np.float32)[roi]
-    zeros9 = np.zeros(n, dtype=np.float32)[roi]
-    ebsd_data = ebsd.create_group("Data")
-    ebsd_data.create_dataset("DD", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("MAD", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("MADPhase", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("NIndexedBands", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("PCX", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("PCY", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("PHI", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("Phase", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("RadonBandCount", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("RadonQuality", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("RawPatterns", data=s.data.reshape((n, sy, sx))[roi])
-    ebsd_data.create_dataset("X BEAM", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("X SAMPLE", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("Y BEAM", dtype=np.int32, data=ones9)
-    ebsd_data.create_dataset("Y SAMPLE", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("Z SAMPLE", dtype=np.float32, data=zeros9)
-    ebsd_data.create_dataset("phi1", dtype=np.float32, data=ones9)
-    ebsd_data.create_dataset("phi2", dtype=np.float32, data=ones9)
-
-    # Header
-    ebsd_header = ebsd.create_group("Header")
-    ebsd_header.create_dataset("CameraTilt", dtype=float, data=0)
-    ebsd_header.create_dataset("DetectorFullHeightMicrons", dtype=np.int32, data=sy)
-    ebsd_header.create_dataset("DetectorFullWidthMicrons", dtype=np.int32, data=sx)
-    grid_type = ebsd_header.create_dataset("Grid Type", shape=(1,), dtype="|S9")
-    grid_type[()] = b"isometric"
-    # ebsd_header.create_dataset("KV", dtype=float, data=20)
-    ebsd_header.create_dataset("MADMax", dtype=float, data=1.5)
-    ebsd_header.create_dataset("Magnification", dtype=float, data=200)
-    ebsd_header.create_dataset("MapStepFactor", dtype=float, data=4)
-    ebsd_header.create_dataset("MaxRadonBandCount", dtype=np.int32, data=11)
-    ebsd_header.create_dataset("MinIndexedBands", dtype=np.int32, data=5)
-    ebsd_header.create_dataset("NCOLS", dtype=np.int32, data=nx)
-    ebsd_header.create_dataset("NROWS", dtype=np.int32, data=ny)
-    ebsd_header.create_dataset("NPoints", dtype=np.int32, data=n)
-    original_file = ebsd_header.create_dataset("OriginalFile", shape=(1,), dtype="|S50")
-    original_file[()] = b"/a/home/for/your/data.h5"
-    ebsd_header.create_dataset("PatternHeight", dtype=np.int32, data=sy)
-    ebsd_header.create_dataset("PatternWidth", dtype=np.int32, data=sx)
-    ebsd_header.create_dataset("PixelByteCount", dtype=np.int32, data=1)
-    s_mean = s.nanmean((2, 3)).data.astype(np.uint8)
-    ebsd_header.create_dataset("SEM Image", data=skc.gray2rgb(s_mean))
-    ebsd_header.create_dataset("SEPixelSizeX", dtype=float, data=1)
-    ebsd_header.create_dataset("SEPixelSizeY", dtype=float, data=1)
-    ebsd_header.create_dataset("SampleTilt", dtype=float, data=70)
-    ebsd_header.create_dataset(
-        "StaticBackground", dtype=np.uint16, data=s.static_background
-    )
-    ebsd_header.create_dataset("TopClip", dtype=float, data=1)
-    ebsd_header.create_dataset("UnClippedPatternHeight", dtype=np.int32, data=sy)
-    ebsd_header.create_dataset("WD", dtype=float, data=1)
-    ebsd_header.create_dataset("XSTEP", dtype=float, data=1.5)
-    ebsd_header.create_dataset("YSTEP", dtype=float, data=1.5)
-    ebsd_header.create_dataset("ZOffset", dtype=float, data=0)
-    # Phases
-    phase = ebsd_header.create_group("Phases/1")
-    formula = phase.create_dataset("Formula", shape=(1,), dtype="|S2")
-    formula[()] = b"Ni"
-    phase.create_dataset("IT", dtype=np.int32, data=225)
-    phase.create_dataset(
-        "LatticeConstants",
-        dtype=np.float32,
-        data=np.array([3.56, 3.56, 3.56, 90, 90, 90]),
-    )
-    name = phase.create_dataset("Name", shape=(1,), dtype="|S6")
-    name[()] = b"Nickel"
-    phase.create_dataset("Setting", dtype=np.int32, data=1)
-    space_group = phase.create_dataset("SpaceGroup", shape=(1,), dtype="|S5")
-    space_group[()] = b"Fm-3m"
-    atom_pos = phase.create_group("AtomPositions")
-    atom_pos1 = atom_pos.create_dataset("1", shape=(1,), dtype="|S17")
-    atom_pos1[()] = b"Ni,0,0,0,1,0.0035"
-
-    # SEM
-    sem = ebsd.create_group("SEM")
-    sem.create_dataset("IX", dtype=np.int32, data=ix)
-    sem.create_dataset("IY", dtype=np.int32, data=iy)
-    sem.create_dataset("ZOffset", dtype=float, data=0)
-
-    f.close()
-
-    yield fpath
+    path = tmpdir / "patterns_roi_nonrectangular.h5"
+    create_dummy_bruker_h5ebsd_nonrectangular_roi_file(path)
+    yield path